services/storage_account/func_get_subscription_list/requirements.txt
azure-functions
azure-mgmt-subscription
azure-mgmt-resource
azure.identity
azure.data.tables
azure-keyvault-secrets
black
pandas
pytest
pytest-cov
python-dotenv